feat(config): temporary force convert `ACCESS_TOKEN_EXPIRE_MINUTES` type to int

+++ api/config.py
@@ -1,3 +1,4 @@
+# TODO: Refactor whole config file
 import os
 import sys
 
@@ -52,5 +53,7 @@ class Config:
 
     SECRET_KEY = _getenv("SECRET_KEY", "m3ow87")
     # 60 minutes * 24 hours * 8 days = 8 days
-    ACCESS_TOKEN_EXPIRE_MINUTES = _getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 60 * 24 * 8)
+    ACCESS_TOKEN_EXPIRE_MINUTES = int(
+        _getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 60 * 24 * 8)
+    )
     ACCESS_TOKEN_ALGORITHM = _getenv("ACCESS_TOKEN_ALGORITHM", "HS256")
